Refactor unit tests for tasks/receive tasks
